requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
cryptography==40.0.2
fastapi==0.95.2
filetype==1.2.0
flatbuffers==23.5.9
frozenlist==1.3.3
h11==0.14.0
huggingface==0.0.1
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
llvmlite==0.40.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
onnxruntime==1.14.1
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
pooch==1.7.0
protobuf==4.23.1
psutil==5.9.5
pycparser==2.21
pydantic==1.10.7
PyMatting==1.1.8
python-multipart==0.0.6
PyWavelets==1.4.1
rembg==2.0.37
requests==2.30.0
scikit-image==0.19.3
scipy==1.10.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tifffile==2023.4.12
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.15.0
